+++ CHANGELOG.md
@@ -2,10 +2,14 @@
 
 ## v1.0.0-beta.2
 
-### Removed
+### Changed
 
-- Removed the now redundant method option from backend plugin actions ([#443](https://github.com/starfederation/datastar/issues/443)).
+- Improved the handling of invalid expressions and signals, and made error handling generally more granular ([#452](https://github.com/starfederation/datastar/issues/452)).
 
 ### Fixed
 
 - Fixed dashes not being trimmed from keys when double dashes were used ([#450](https://github.com/starfederation/datastar/issues/450)).
+
+### Removed
+
+- Removed the now redundant method option from backend plugin actions ([#443](https://github.com/starfederation/datastar/issues/443)).